chore: redo project structure
